restaurant card, country picker and small fix
